selecttest/Selection-collapseToEnd.html

<!DOCTYPE html>
<title>The collapseToEnd method</title>
<div id=log></div>
<script src=http://w3c-test.org/resources/testharness.js></script>
<script>
test(function() {
  var sel = getSelection();
  sel.removeAllRanges();
  assert_throws("INVALID_STATE_ERR", function() { sel.collapseToEnd(); })
});
</script>
